requirements.txt
# === CORE FRAMEWORK ===
fastapi==0.104.1              # API moderna y async
uvicorn[standard]==0.24.0     # ASGI server con performance optimizado
celery[redis]==5.3.4          # Task queue con Redis
redis==5.0.1                  # Cache y message broker

# === DATABASE ===
sqlalchemy==2.0.23            # ORM moderno con type hints
alembic==1.12.1              # Database migrations
psycopg2-binary==2.9.9       # PostgreSQL adapter optimizado
asyncpg==0.29.0              # Async PostgreSQL para FastAPI

# === VALIDATION & SERIALIZATION ===
pydantic==2.5.0              # Data validation con V2
pydantic-settings==2.1.0     # Settings management
email-validator==2.1.0       # Email validation
python-dotenv==1.0.0         # Load environment variables

# === WHATSAPP & API CLIENTS ===
httpx==0.25.2                # Async HTTP client
aiohttp==3.9.1               # Alternative async HTTP
requests==2.31.0             # Sync HTTP fallback

# === AUDIO PROCESSING ===
vosk==0.3.45                 # Speech recognition
ffmpeg-python==0.2.0         # Audio conversion
wave                         # Built-in wave processing

# === AI & ML ===
google-generativeai>=0.8.5   # Gemini AI (updated for latest models)
numpy==1.24.3                # Scientific computing
pandas==2.0.3                # Data analysis
google-api-core==2.15.0      # Google API core
google-auth==2.25.2          # Google authentication

# === MONITORING & LOGGING ===
structlog==23.2.0            # Structured logging
sentry-sdk==1.39.2           # Error tracking
prometheus-client==0.19.0    # Metrics collection

# === SECURITY ===
cryptography==41.0.8         # Encryption utilities
python-jose[cryptography]==3.3.0  # JWT handling
passlib[bcrypt]==1.7.4       # Password hashing

# === ENVIRONMENT & CONFIG ===
python-dotenv==1.0.0         # Environment variables
pyyaml==6.0.1                # YAML config files

# === TESTING (DEV) ===
pytest==7.4.3               # Testing framework
pytest-asyncio==0.21.1      # Async testing
pytest-cov==4.1.0           # Coverage reports
httpx==0.25.2               # API testing
factory-boy==3.3.0          # Test data factories

# === DEVELOPMENT TOOLS ===
black==23.11.0               # Code formatting
isort==5.12.0                # Import sorting
flake8==6.1.0                # Linting
mypy==1.7.1                  # Type checking

# === DOCUMENTATION ===
sphinx==7.2.6               # Documentation generation
sphinx-rtd-theme==1.3.0     # ReadTheDocs theme

# === DEPLOYMENT ===
gunicorn==21.2.0             # WSGI server
supervisor==4.2.5            # Process management
nginx-python==1.1.0          # Nginx config helpers

# === DASHBOARD & VISUALIZATION ===
streamlit==1.28.2           # Dashboard framework
plotly==5.17.0              # Interactive charts
pandas==2.1.4               # Data manipulation
numpy==1.26.2               # Numerical computing

# === UTILITIES ===
click==8.1.7                 # CLI framework
rich==13.7.0                 # Rich terminal output
typer==0.9.0                 # Modern CLI creation
schedule==1.2.0             # Job scheduling